feat: Refactor into cross-platform architecture with core monitoring logic, platform-specific app implementations, and updated readme.
