+++ app_linux.py
@@ -1,0 +1,9 @@
+"""Monarx - Linux system tray implementation."""
+
+
+
+def run():
+    """Run the Linux app."""
+    print("Linux support coming soon!")
+    print("Will use pystray for system tray integration.")
+    raise NotImplementedError("Linux support not yet implemented")
