+++ app_windows.py
@@ -1,0 +1,7 @@
+"""Monarx - Windows system tray implementation."""
+
+def run():
+    """Run the Windows app."""
+    print("Windows support coming soon!")
+    print("Will use pystray for system tray integration.")
+    raise NotImplementedError("Windows support not yet implemented")
